requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.9.2
sqlalchemy==2.0.35
aiosmtplib==3.0.2
python-multipart==0.0.12
slowapi==0.1.9
python-dotenv==1.0.1
